moved givertype and recipienttype to overviewMacro
